Column group definitions

Creates a column group definition for grouping columns in a Cheetah Grid widget.

Usage

column_group(name = NULL, columns, header_style = NULL)

Arguments

name

Character string. The name to display for the column group.

columns

Character vector. The names of the columns to include in this group.

header_style

Named list of possibleCSS style properties to apply to the column group header.

Value

A list containing the column group definition.

Examples

cheetah(
  iris,
  columns = list(
    Sepal.Length = column_def(name = "Length"),
    Sepal.Width = column_def(name = "Width"),
    Petal.Length = column_def(name = "Length"),
    Petal.Width = column_def(name = "Width")
  ),
  column_group = list(
    column_group(name = "Sepal", columns = c("Sepal.Length", "Sepal.Width")),
    column_group(name = "Petal", columns = c("Petal.Length", "Petal.Width"))
  )
)