Create a Cheetah Grid widget

Creates a high-performance table widget using Cheetah Grid.

Usage

cheetah(
  data,
  columns = NULL,
  column_group = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  rownames = TRUE,
  search = c("disabled", "exact", "contains"),
  sortable = TRUE
)

Arguments

data

A data frame or matrix to display

columns

A list of column definitions. Each column can be customized using column_def().

column_group

A list of column groups. Each group can be customized using

width

Width of the widget

height

Height of the widget

elementId

The element ID for the widget

rownames

Logical. Whether to show rownames. Defaults to TRUE.

search

Whether to enable a search field on top of the table. Default to disabled. Use exact for exact matching or contains to get larger matches.

sortable

Logical. Whether to enable sorting on all columns. Defaults to TRUE.

Value

An HTML widget object of class 'cheetah' that can be:

Rendered in R Markdown documents

Used in Shiny applications

Displayed in R interactive sessions

The widget renders as an HTML table with all specified customizations.